ai-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
numpy==1.24.3  # Versão compatível com TensorFlow
tensorflow==2.10.0
opencv-python-headless==4.8.1.78
python-multipart==0.0.6